refactor(server): extract supporting CLI enums to cli/enums.rs
FU-4 of v0.8.0 refinement follow-up. The `cli.rs` file (1972 LOC)
was dominated by the giant `Commands` enum (clap subcommands), but
~210 LOC at the top were 7 supporting enums (`ValidationTarget`,
`SyncSource`, `DependencyDirection`, `AgentBackend`,
`CreateSkillMethod`, `CertAction`, `OutputFormat`).

Convert `cli.rs` to `cli/` directory module; move the 7 enums to
`cli/enums.rs` with `pub use enums::*` in `mod.rs` so external
callers (`crate::cli::OutputFormat`, etc.) keep working unchanged.

cli/mod.rs:    1972 -> 1769 LOC (-10%)
cli/enums.rs:  220 LOC (new, with module docstring)

The `Commands` enum stays in mod.rs because splitting clap
variants across files fragments a coherent CLI surface. This
remains the right call.

480 server-lib tests pass.
